moved test to proper package
